+++ test/Fixture/Methods/FinalInAbstractClassRule/Success/AbstractClassWithProtectedMethodAndOrmEntityAnnotationInMultilineDocBlock.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ergebnis\PHPStan\Rules\Test\Fixture\Methods\FinalInAbstractClassRule\Success;
+
+/**
+ * @ORM\Entity
+ */
+abstract class AbstractClassWithProtectedMethodAndOrmEntityAnnotationInMultilineDocBlock
+{
+    protected function method(): void
+    {
+    }
+}