+++ test/Fixture/Methods/FinalInAbstractClassRule/Success/AbstractClassWithPublicMethodAndEmbeddableAnnotationInInlineDocBlock.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ergebnis\PHPStan\Rules\Test\Fixture\Methods\FinalInAbstractClassRule\Success;
+
+/** @Embeddable */
+abstract class AbstractClassWithPublicMethodAndEmbeddableAnnotationInInlineDocBlock
+{
+    public function method(): void
+    {
+    }
+}