+++ test/Integration/Closures/NoParameterWithNullableTypeDeclarationRuleTest.php
@@ -48,6 +48,10 @@ final class NoParameterWithNullableTypeDeclarationRuleTest extends Testing\RuleT
                     'Closure has parameter $bar with a nullable type declaration.',
                     26,
                 ],
+                [
+                    'Closure has parameter $bar with a nullable type declaration.',
+                    30,
+                ],
             ],
         );
     }