+++ test/Integration/Methods/NoNullableReturnTypeDeclarationRuleTest.php
@@ -76,6 +76,10 @@ final class NoNullableReturnTypeDeclarationRuleTest extends Testing\RuleTestCase
                     'Method toString() in anonymous class has a nullable return type declaration.',
                     36,
                 ],
+                [
+                    'Method toString() in anonymous class has a nullable return type declaration.',
+                    43,
+                ],
             ],
         );
     }