+++ test/Integration/Methods/NoParameterWithNullableTypeDeclarationRuleTest.php
@@ -76,6 +76,10 @@ final class NoParameterWithNullableTypeDeclarationRuleTest extends Testing\RuleT
                     'Method foo() in anonymous class has parameter $bar with a nullable type declaration.',
                     42,
                 ],
+                [
+                    'Method foo() in anonymous class has parameter $bar with a nullable type declaration.',
+                    49,
+                ],
             ],
         );
     }