+++ CHANGELOG.md
@@ -14,7 +14,7 @@ For a full diff see [`2.5.0...main`][2.5.0...main].
 - Adjusted `Methods\FinalInAbstractClassRule` to ignore Doctrine embeddables and entities ([#396]), by [@localheinz]
 - Adjusted `Expressions\NoCompactRule` to detect usages of `compact()` with incorrect case ([#889]), by [@localheinz]
 - Adjusted `Methods\PrivateInFinalClass` to use more appropriate message when detecting a `protected` method in an anonymous class ([#890]), by [@localheinz]
-- Adjusted `Methods\PrivateInFinalClass` to ignore `protected` methods from traits([#891]), by [@localheinz]
+- Adjusted `Methods\PrivateInFinalClass` to ignore `protected` methods from traits ([#891]), by [@localheinz]
 
 ## [`2.5.0`][2.5.0]